Convert FormulaAuditor test to spec.
